coversion by platform configuration moved to cmd
