+++ cmd/platform/process.go
@@ -1,6 +1,7 @@
 package platform
 
 import (
+	"cloud-mta-build-tool/mta"
 	"gopkg.in/yaml.v2"
 
 	"cloud-mta-build-tool/cmd/logs"
@@ -13,4 +14,31 @@ func Parse(data []byte) Platforms {
 		logs.Logger.Error("Yaml file is not valid, Error: " + err.Error())
 	}
 	return platforms
+}
+
+func ConvertTypes(iCfg mta.MTA, eCfg Platforms, targetPlatform string) {
+	// todo get from config
+	const (
+		SchemaVersion = "3.1"
+	)
+	tpl := platformConfig(eCfg, targetPlatform)
+	for i, v := range iCfg.Modules {
+		*iCfg.SchemaVersion = SchemaVersion
+		for _, em := range tpl.Modules {
+			if v.Type == em.NativeType {
+				iCfg.Modules[i].Type = em.PlatformType
+			}
+		}
+	}
+}
+
+func platformConfig(eCfg Platforms, targetPlatform string) Modules {
+	var tpl Modules
+	for _, tp := range eCfg.Platforms {
+		if tp.Name == targetPlatform {
+			tpl.Name = tp.Name
+			tpl.Modules = tp.Modules
+		}
+	}
+	return tpl
 }
