+++ mta/desc.go
@@ -6,8 +6,6 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
-
-	"cloud-mta-build-tool/cmd/platform"
 )
 
 // The deployment descriptor shall be located within the META-INF folder of the JAR.
@@ -61,14 +59,15 @@ func printToFile(file io.Writer, mtaStr *Modules) {
 	fmt.Fprint(file, contentType+applicationZip)
 }
 
-func GenMtad(mtaStr MTA, targetPath string) error {
+func GenMtad(mtaStr MTA, targetPath string, convertTypes func(mtaStr MTA)) error {
 	// Create META-INF folder under the mtar folder
 	createDirIfNotExist(filepath.Join(targetPath, metaInf))
-	// Load platform configuration file
-	platformCfg := platform.Parse(platform.PlatformConfig)
-	// Modify MTAD object according to platform types
-	// Todo platform should provided as command parameter
-	ConvertTypes(mtaStr, platformCfg, "cf")
+	convertTypes(mtaStr)
+	//// Load platform configuration file
+	//platformCfg := platform.Parse(platform.PlatformConfig)
+	//// Modify MTAD object according to platform types
+	//// Todo platform should provided as command parameter
+	//ConvertTypes(mtaStr, platformCfg, "cf")
 	// Create readable Yaml before writing to file
 	mtad, err := Marshal(mtaStr)
 	// Write back the MTAD to the META-INF folder
@@ -78,8 +77,8 @@ func GenMtad(mtaStr MTA, targetPath string) error {
 	return err
 }
 
-func GenMetaInfo(tmpDir string, mtaStr MTA, modules []string) error {
-	err := GenMtad(mtaStr, tmpDir)
+func GenMetaInfo(tmpDir string, mtaStr MTA, modules []string, convertTypes func(mtaStr MTA)) error {
+	err := GenMtad(mtaStr, tmpDir, convertTypes)
 	if err == nil {
 		// Create MANIFEST.MF file
 		file, _ := createFile(filepath.Join(tmpDir, metaInf, manifest))
